+++ NEWS.md
@@ -2,6 +2,10 @@
 
 * associated Journal of Open Source Software paper now published and should be used as reference.
 
+* `calibrate_LLR()` changed argument position to make it more compatible with the tidy approach. Now the data to calibrate is the first argument while the calibration dataset is the second argument.
+
+* The calibration dataset is now optional for `calibrate_LLR()`. If absent, the calibration is done leave-one-out. This is useful to see the LLRs when testing a method on a training dataset.
+
 * `contentmask()` also accepts sentence-tokenised corpora (the outputs of `tokenize_sents()`) as input, this option also allowing parallel processing.
 
 * `chunk_texts()` now outputs texts that keep the same spaces present in the original and no longer outputs texts with spaces around the punctuation marks.
@@ -10,7 +14,7 @@
 
 * `vectorize()` and two functions that call it (`delta()` and `ngram_tracing()`) now have a new argument called 'cross_boundaries'. If FALSE, n-grams do not cross sentence boundaries (which was the default behaviour in previous versions). This change simply means that the user can now choose to cross sentence boundaries when making n-grams if they wish. The behaviour of these functions is therefore also now clearer.
 
-* the progres bar is now optional for all authorship analysis functions (but default is set to TRUE).
+* the progress bar is now optional for all authorship analysis functions (but default is set to TRUE).
 
 # idiolect 1.2.0